remove unused explicit; move cycle constructor to cpp; inverse direction
